+++ css/style.css
@@ -1416,3 +1416,7 @@ main {
 .hidden {
   display: none;
 }
+
+.pristine-error {
+  color:#ee8e20;
+}
