+++ index.html
@@ -5,6 +5,7 @@
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
+  <link rel="stylesheet" href="nouislider/nouislider.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
 </head>
@@ -58,7 +59,7 @@
                   <!-- Изменение глубины эффекта, накладываемого на изображение -->
                   <fieldset class="img-upload__effect-level  effect-level">
                     <input class="effect-level__value" type="number" step="any" name="effect-level" value="">
-                    <div class="effect-level__slider"></div>
+                    <div class="effect-level__slider" id="slider"></div>
                   </fieldset>
 
                   <!-- Кнопка для закрытия формы редактирования изображения -->
@@ -115,7 +116,9 @@
 
                 <!-- Добавление хэш-тегов и комментария к изображению -->
                 <fieldset class="img-upload__text text">
+                  <div class="img-upload__error-text">
                   <input class="text__hashtags" name="hashtags" placeholder="#ХэшТег">
+                  </div>
                   <textarea class="text__description" name="description" placeholder="Ваш комментарий..." maxlength="140"
                     data-pristine-maxlength-message="Максимальная длина 140 символов"></textarea>
                 </fieldset>
@@ -143,7 +146,12 @@
         </div>
 
         <!-- Информация об изображении. Подпись, комментарии, количество лайков -->
-        <div class="big-picture__social  social">
+        <div class="big-picture__social  social"
+
+
+
+
+        >
           <div class="social__header">
             <img class="social__picture" src="img/avatar-1.svg" alt="Аватар автора фотографии" width="35" height="35">
             <p class="social__caption">Тестим новую камеру! =)</p>
@@ -231,6 +239,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script src="pristine/pristine.min.js"></script>
+<script src="nouislider/nouislider.js"></script>
 <script src = "js/main.js" type = "module"></script>
 </body>
 </html>
